Move config module out of common
- Moves the `/src/common/config` to `/src/config`
- The config module should be seen as something horizontal to the rest of the application (accessible from domain, data and routing layers)
- The common module should be seen as common utilities used on the route layer
