+++ src/config/entities/configuration.ts
