+++ src/datasources/config-api/config-api.service.ts
@@ -1,7 +1,7 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { Page } from '../../domain/entities/page.entity';
 import { Chain } from '../../domain/chains/entities/chain.entity';
-import { IConfigurationService } from '../../common/config/configuration.service.interface';
+import { IConfigurationService } from '../../config/configuration.service.interface';
 import { CacheFirstDataSource } from '../cache/cache.first.data.source';
 import { IConfigApi } from '../../domain/interfaces/config-api.interface';
