+++ src/routes/balances/balances.controller.spec.ts
@@ -11,7 +11,7 @@ import { BalancesModule } from './balances.module';
 import {
   fakeConfigurationService,
   TestConfigurationModule,
-} from '../../common/config/__tests__/test.configuration.module';
+} from '../../config/__tests__/test.configuration.module';
 import exchangeFiatCodesFactory from '../../domain/exchange/entities/__tests__/fiat-codes.factory';
 import {
   fakeCacheService,
