+++ src/routes/cache-hooks/cache-hooks.controller.spec.ts
@@ -2,7 +2,7 @@ import { INestApplication } from '@nestjs/common';
 import {
   fakeConfigurationService,
   TestConfigurationModule,
-} from '../../common/config/__tests__/test.configuration.module';
+} from '../../config/__tests__/test.configuration.module';
 import {
   fakeCacheService,
   TestCacheModule,
